refactor(orchestrator): extract StreamNormalizationMixin from the God-file (D-164 Slice 1)
Move _normalize_stream_event / _normalize_ui_component_event / _sanitize_error_for_user /
_recover_structured_event (+ their two event-type frozensets) verbatim into a cohesive
StreamNormalizationMixin (SRP). OrchestratorClient(ProbabilityTutorBrain,
StreamNormalizationMixin) — brain stays first in the MRO; every self._x resolves through it,
so behaviour is byte-identical. Append the mixin to TUTOR_SOURCE_FILES (one line) so the
persistence-authority source-inspection test auto-follows via read_tutor_source(). Drop the
now-orphaned imports. God-file 3,832 -> 3,678 lines.
