+++ app/infrastructure/clients/orchestrator/tutor_sources.py
@@ -25,6 +25,7 @@ _ROOT = Path(__file__).resolve().parents[4]
 TUTOR_SOURCE_FILES: tuple[str, ...] = (
     "app/infrastructure/clients/orchestrator_client.py",
     "app/services/skills/probability_tutor_brain.py",
+    "app/infrastructure/clients/orchestrator/stream_normalization.py",  # D-164 Slice 1
 )
